[#20241026] Split domain logic classes into discrete files
Enhances maintainability by isolating logic components.

Relates to project refactoring objectives.
